[install-sh] Modified gdrive install

+++ install.sh
@@ -158,9 +158,11 @@ DOTFILES_DIRECTORY=$HOME/.dotfiles
             fi
 
             # gdrive
-            green-echo "install gdrive"
-            wget -q "https://docs.google.com/uc?id=0B3X9GlR6EmbnQ0FtZmJJUXEyRTA&export=download" -O $HOME/.local/bin/gdrive
-            chmod +x $HOME/.local/bin/gdrive
+            if [ ! -f $HOME/.local/bin/gdrive ]; then
+                green-echo "install gdrive"
+                wget -q "https://docs.google.com/uc?id=0B3X9GlR6EmbnQ0FtZmJJUXEyRTA&export=download" -O $HOME/.local/bin/gdrive
+                chmod +x $HOME/.local/bin/gdrive
+            fi
             ;;
         darwin*)
             # install from brew
